requirements.txt
# 依赖列表

# 核心依赖
requests>=2.31.0
PyYAML>=6.0
loguru>=0.7.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 可选依赖
# matplotlib>=3.7.0  # 用于可视化
# jupyter>=1.0.0     # 用于交互式分析
